pyproject.toml
[tool.poetry]
name = "Graha Tia"
version = "0.0.1"
description = "A helper bot for organizing your FFXIV needs."
authors = ["Alex Nørgaard <umbra@abstractumbra.dev>"]

[tool.poetry.dependencies]
python = "^3.12"
"discord.py" = { git = "https://github.com/Rapptz/discord.py.git", extras = [
    "speed",
], rev = "master" }
discord-ext-menus = { git = "https://github.com/Rapptz/discord-ext-menus", rev = "master" }
jishaku = "*"
"mystbin.py" = "*"
asyncpg = "*"
sentry-sdk = { version = "*", extras = ["asyncpg", "aiohttp"] }
lru-dict = "*"
parsedatetime = "*"
python-dateutil = "*"
psutil = "*"
pygit2 = "*"
beautifulsoup4 = "*"
lxml = "*"

[tool.poetry.group.dev.dependencies]
ruff = "*"
typing-extensions = "*"
asyncpg-stubs = "*"
"pre-commit" = "*"

[tool.poetry.group.speed.dependencies]
orjson = "*"
uvloop = "*"

[tool.ruff]
line-length = 125
target-version = "py312"

[tool.ruff.lint]
select = [
    "A",
    "ANN",
    "ASYNC",
    "B",
    "BLE",
    "C4",
    "COM",
    "DTZ",
    "E",
    "EM",
    "ERA",
    "F",
    "FA",
    "FBT",
    "FURB",
    "G",
    "I",
    "INP",
    "ISC",
    "NPY",
    "PD",
    "PERF",
    "PGH",
    "PIE",
    "PLC",
    "PLE",
    "PLW",
    "PTH",
    "PYI",
    "Q",
    "Q003",
    "RET",
    "RSE",
    "RUF",
    "S",
    "SIM",
    "SLOT",
    "T20",
    "TC",
    "TID",
    "TRY",
    "UP",
    "YTT",
    # "DOC", # not in a bot, please
]
ignore = [
    "ANN401",
    "ASYNC109", # dpy commonly uses this parameter
    "EM101",    # in this case this is okay
    "F401",
    "F402",
    "F403",
    "F405",
    "PD011",    # this is not a numpy codebase
    "PERF203",
    "PLC0414",  # pyright ruling for `as` imports needed
    "Q000",
    "RUF001",
    "RUF009",
    "S101",     # assert is used for type checking in this codebase
    "SIM105",
    "TRY003",   # over-eager rule
    "TRY301",   # unrealistic rule
    "UP034",
    "UP038",
]
unfixable = [
    "E501", # line length handled in other ways by ruff format
    "ERA",  # Don't delete commented out code
]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint.isort]
split-on-trailing-comma = true
combine-as-imports = true

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false
parametrize-names-type = "csv"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.lint.flake8-tidy-imports.banned-api]
# https://discuss.python.org/t/problems-with-typeis/55410/6
# https://discuss.python.org/t/problems-with-typeis/55410/46
# Until what can go into a TypeIs/TypeGuard changes, these are just dangerous.
"typing.TypeIs".msg = "TypeIs is fundamentally unsafe, even when using it as described to be safe"
"typing.TypeGuard".msg = "TypeGuard is fundamentally unsafe"
"typing_extensions.TypeIs".msg = "TypeIs is fundamentally unsafe, even when using it as described to be safe"
"typing_extensions.TypeGuard".msg = "TypeGuard is fundamentally unsafe"

[tool.pyright]
include = ["bot.py", "utilities/**/*.py", "extensions/**/*.py"]
venv = ".venv"
venvPath = "."
useLibraryCodeForTypes = true
typeCheckingMode = "basic"
pythonVersion = "3.12"
pythonPlatform = "All"
strictListInference = true
strictDictionaryInference = true
strictSetInference = true
strictParameterNoneValue = true
reportMissingImports = "error"
reportUnusedImport = "error"
reportUnusedClass = "error"
reportUnusedFunction = "error"
reportUnusedVariable = "error"
reportGeneralTypeIssues = "error"
reportDuplicateImport = "error"
reportUntypedFunctionDecorator = "error"
reportUnnecessaryTypeIgnoreComment = "warning"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
